- tauri.conf.json: AxInvest + v2.0.0
- locale JSON: 上游重建 + AxInvest stockAnalysis 键还原
- i18n allowlist: 采用上游 269 条目
- gateway: Rust 2024 edition 兼容
- SubAgentCard/SettingsSidebar: 合并双方改动

+++ src-tauri/crates/core/src/entity/gateway_request_logs.rs
@@ -14,8 +14,8 @@ pub struct Model {
     pub provider_id: Option<String>,
     pub status_code: i32,
     pub duration_ms: i32,
-    pub request_tokens: i32,
-    pub response_tokens: i32,
+    pub request_tokens: i64,
+    pub response_tokens: i64,
     pub error_message: Option<String>,
     pub created_at: i64,
 }
