+++ src-tauri/crates/core/src/repo/gateway.rs
@@ -97,8 +97,8 @@ pub async fn record_usage(
         key_id: Set(key_id.to_string()),
         provider_id: Set(provider_id.to_string()),
         model_id: Set(model_id.map(|s| s.to_string())),
-        request_tokens: Set(request_tokens as i32),
-        response_tokens: Set(response_tokens as i32),
+        request_tokens: Set(request_tokens as i64),
+        response_tokens: Set(response_tokens as i64),
         created_at: Set(now_ts()),
         ..Default::default()
     }
